+++ examples/graphics_i2c_72x40.rs
@@ -65,8 +65,8 @@ fn main() -> ! {
     );
 
     let interface = I2CDIBuilder::new().init(i2c);
-    let mut disp: GraphicsMode<_> = Builder::new()
-        .size(DisplaySize::Display72x40)
+    let mut disp: GraphicsMode<_, _> = Builder::new()
+        .size(DisplaySize72x40)
         .connect(interface)
         .into();
     disp.init().unwrap();