+++ src/mode/terminal.rs
@@ -31,7 +31,7 @@ use crate::{
     brightness::Brightness,
     command::AddrMode,
     displayrotation::DisplayRotation,
-    displaysize::DisplaySize,
+    displaysize::*,
     mode::{
         displaymode::DisplayModeTrait,
         terminal::TerminalModeError::{InterfaceError, OutOfBounds, Uninitialized},
@@ -39,6 +39,33 @@ use crate::{
     properties::DisplayProperties,
 };
 use core::{cmp::min, fmt};
+
+/// Extends the [`DisplaySize`](../../displaysize/trait.DisplaySize.html) trait
+/// to include number of characters that can fit on the display.
+pub trait TerminalDisplaySize: DisplaySize {
+    /// The number of characters that can fit on the display at once (w * h / (8 * 8))
+    const CHAR_NUM: u8;
+}
+
+impl TerminalDisplaySize for DisplaySize128x64 {
+    const CHAR_NUM: u8 = 128;
+}
+
+impl TerminalDisplaySize for DisplaySize128x32 {
+    const CHAR_NUM: u8 = 64;
+}
+
+impl TerminalDisplaySize for DisplaySize96x16 {
+    const CHAR_NUM: u8 = 24;
+}
+
+impl TerminalDisplaySize for DisplaySize72x40 {
+    const CHAR_NUM: u8 = 45;
+}
+
+impl TerminalDisplaySize for DisplaySize64x48 {
+    const CHAR_NUM: u8 = 48;
+}
 
 /// Contains the new row that the cursor has wrapped around to
 struct CursorWrapEvent(u8);
@@ -135,17 +162,21 @@ impl<T> IntoTerminalModeResult<T> for Result<T, DisplayError> {
 
 // TODO: Add to prelude
 /// Terminal mode handler
-pub struct TerminalMode<DI> {
-    properties: DisplayProperties<DI>,
+pub struct TerminalMode<DI, DSIZE = DisplaySize128x64>
+where
+    DSIZE: TerminalDisplaySize,
+{
+    properties: DisplayProperties<DI, DSIZE>,
     cursor: Option<Cursor>,
 }
 
-impl<DI> DisplayModeTrait<DI> for TerminalMode<DI>
+impl<DI, DSIZE> DisplayModeTrait<DI, DSIZE> for TerminalMode<DI, DSIZE>
 where
     DI: WriteOnlyDataCommand,
+    DSIZE: TerminalDisplaySize,
 {
     /// Create new TerminalMode instance
-    fn new(properties: DisplayProperties<DI>) -> Self {
+    fn new(properties: DisplayProperties<DI, DSIZE>) -> Self {
         TerminalMode {
             properties,
             cursor: None,
@@ -153,47 +184,34 @@ where
     }
 
     /// Release display interface used by `TerminalMode`
-    fn into_properties(self) -> DisplayProperties<DI> {
+    fn into_properties(self) -> DisplayProperties<DI, DSIZE> {
         self.properties
     }
 }
 
-impl<DI> TerminalMode<DI>
+impl<DI, DSIZE> TerminalMode<DI, DSIZE>
 where
     DI: WriteOnlyDataCommand,
+    DSIZE: TerminalDisplaySize,
 {
     /// Clear the display and reset the cursor to the top left corner
     pub fn clear(&mut self) -> Result<(), TerminalModeError> {
-        let display_size = self.properties.get_size();
-
-        // The number of characters that can fit on the display at once (w * h / 8 * 8)
-        // TODO: Use `display_size.dimensions()`
-        let numchars = match display_size {
-            DisplaySize::Display128x64 => 128,
-            DisplaySize::Display128x32 => 64,
-            DisplaySize::Display96x16 => 24,
-            DisplaySize::Display72x40 => 45,
-            DisplaySize::Display64x48 => 48,
-        };
-
         // Let the chip handle line wrapping so we can fill the screen with blanks faster
         self.properties
             .change_mode(AddrMode::Horizontal)
             .terminal_err()?;
-        let (display_width, display_height) = self.properties.get_dimensions();
-        let (display_x_offset, display_y_offset) = self.properties.display_offset;
         self.properties
             .set_draw_area(
-                (display_x_offset, display_y_offset),
+                (DSIZE::OFFSETX, DSIZE::OFFSETY),
                 (
-                    display_width + display_x_offset,
-                    display_height + display_y_offset,
+                    DSIZE::WIDTH + DSIZE::OFFSETX,
+                    DSIZE::HEIGHT + DSIZE::OFFSETY,
                 ),
             )
             .terminal_err()?;
 
         // Clear the display
-        for _ in 0..numchars {
+        for _ in 0..DSIZE::CHAR_NUM {
             self.properties.draw(&[0; 8]).terminal_err()?;
         }
 
@@ -236,7 +254,7 @@ where
     }
 
     /// Initialise the display in page mode (i.e. a byte walks down a column of 8 pixels) with
-    /// column 0 on the left and column _(display_width - 1)_ on the right, but no automatic line
+    /// column 0 on the left and column _(DSIZE::Width::U8 - 1)_ on the right, but no automatic line
     /// wrapping.
     pub fn init(&mut self) -> Result<(), TerminalModeError> {
         self.properties
@@ -289,14 +307,10 @@ where
 
     /// Reset the draw area and move pointer to the top left corner
     fn reset_pos(&mut self) -> Result<(), TerminalModeError> {
-        let (display_x_offset, display_y_offset) = self.properties.display_offset;
-        self.properties
-            .set_column(display_x_offset)
-            .terminal_err()?;
-        self.properties.set_row(display_y_offset).terminal_err()?;
+        self.properties.set_column(DSIZE::OFFSETX).terminal_err()?;
+        self.properties.set_row(DSIZE::OFFSETY).terminal_err()?;
         // Initialise the counter when we know it's valid
-        let (display_width, display_height) = self.properties.get_dimensions();
-        self.cursor = Some(Cursor::new(display_width, display_height));
+        self.cursor = Some(Cursor::new(DSIZE::WIDTH, DSIZE::HEIGHT));
 
         Ok(())
     }
@@ -415,9 +429,10 @@ where
     }
 }
 
-impl<DI> fmt::Write for TerminalMode<DI>
+impl<DI, DSIZE> fmt::Write for TerminalMode<DI, DSIZE>
 where
     DI: WriteOnlyDataCommand,
+    DSIZE: TerminalDisplaySize,
 {
     fn write_str(&mut self, s: &str) -> Result<(), fmt::Error> {
         s.chars().map(move |c| self.print_char(c)).last();