+++ src/properties.rs
@@ -1,56 +1,51 @@
 //! Container to store and set display properties
-
 use crate::mode::displaymode::DisplayModeTrait;
 use crate::{
     brightness::Brightness,
     command::{AddrMode, Command, VcomhLevel},
     displayrotation::DisplayRotation,
-    displaysize::DisplaySize,
+    displaysize::{DisplaySize, DisplaySize128x64},
 };
 use display_interface::{DataFormat::U8, DisplayError, WriteOnlyDataCommand};
 
 /// Display properties struct
-pub struct DisplayProperties<DI> {
+pub struct DisplayProperties<DI, DSIZE = DisplaySize128x64> {
     iface: DI,
-    display_size: DisplaySize,
     display_rotation: DisplayRotation,
-    pub(crate) display_offset: (u8, u8),
     addr_mode: AddrMode,
+    size: DSIZE,
 }
 
-impl<DI> DisplayProperties<DI> {
+impl<DI, DSIZE> DisplayProperties<DI, DSIZE>
+where
+    DSIZE: DisplaySize,
+{
     /// Create new DisplayProperties instance
     pub fn new(
         iface: DI,
-        display_size: DisplaySize,
+        size: DSIZE,
         display_rotation: DisplayRotation,
-    ) -> DisplayProperties<DI> {
-        let display_offset = match display_size {
-            DisplaySize::Display128x64 => (0, 0),
-            DisplaySize::Display128x32 => (0, 0),
-            DisplaySize::Display96x16 => (0, 0),
-            DisplaySize::Display72x40 => (28, 0),
-            DisplaySize::Display64x48 => (32, 0),
-        };
-
+    ) -> DisplayProperties<DI, DSIZE> {
         DisplayProperties {
             iface,
-            display_size,
             display_rotation,
-            display_offset,
             addr_mode: AddrMode::Page, // reset value
+            size,
         }
     }
+}
 
+impl<DI, DSIZE> DisplayProperties<DI, DSIZE> {
     /// Releases the display interface
     pub fn release(self) -> DI {
         self.iface
     }
 }
 
-impl<DI> DisplayProperties<DI>
+impl<DI, DSIZE> DisplayProperties<DI, DSIZE>
 where
     DI: WriteOnlyDataCommand,
+    DSIZE: DisplaySize,
 {
     /// Initialise the display in column mode (i.e. a byte walks down a column of 8 pixels) with
     /// column 0 on the left and column _(display_width - 1)_ on the right.
@@ -60,29 +55,19 @@ where
 
     /// Initialise the display in one of the available addressing modes
     pub fn init_with_mode(&mut self, mode: AddrMode) -> Result<(), DisplayError> {
-        // TODO: Break up into nice bits so display modes can pick whathever they need
-        let (_, display_height) = self.display_size.dimensions();
-
         let display_rotation = self.display_rotation;
 
         Command::DisplayOn(false).send(&mut self.iface)?;
         Command::DisplayClockDiv(0x8, 0x0).send(&mut self.iface)?;
-        Command::Multiplex(display_height - 1).send(&mut self.iface)?;
+        Command::Multiplex(DSIZE::HEIGHT - 1).send(&mut self.iface)?;
         Command::DisplayOffset(0).send(&mut self.iface)?;
         Command::StartLine(0).send(&mut self.iface)?;
         // TODO: Ability to turn charge pump on/off
         Command::ChargePump(true).send(&mut self.iface)?;
         Command::AddressMode(mode).send(&mut self.iface)?;
 
+        self.size.configure(&mut self.iface)?;
         self.set_rotation(display_rotation)?;
-
-        match self.display_size {
-            DisplaySize::Display128x32 => Command::ComPinConfig(false, false).send(&mut self.iface),
-            DisplaySize::Display128x64 => Command::ComPinConfig(true, false).send(&mut self.iface),
-            DisplaySize::Display96x16 => Command::ComPinConfig(false, false).send(&mut self.iface),
-            DisplaySize::Display72x40 => Command::ComPinConfig(true, false).send(&mut self.iface),
-            DisplaySize::Display64x48 => Command::ComPinConfig(true, false).send(&mut self.iface),
-        }?;
 
         self.set_brightness(Brightness::default())?;
         Command::VcomhDeselect(VcomhLevel::Auto).send(&mut self.iface)?;
@@ -178,11 +163,6 @@ where
             .try_for_each(|c| self.iface.send_data(U8(&c)))
     }
 
-    /// Get the configured display size
-    pub fn get_size(&self) -> DisplaySize {
-        self.display_size
-    }
-
     /// Get display dimensions, taking into account the current rotation of the display
     ///
     /// ```rust
@@ -192,7 +172,7 @@ where
     /// #
     /// let disp = DisplayProperties::new(
     ///     interface,
-    ///     DisplaySize::Display128x64,
+    ///     DisplaySize128x64,
     ///     DisplayRotation::Rotate0,
     /// );
     /// assert_eq!(disp.get_dimensions(), (128, 64));
@@ -200,17 +180,15 @@ where
     /// # let interface = StubInterface;
     /// let rotated_disp = DisplayProperties::new(
     ///     interface,
-    ///     DisplaySize::Display128x64,
+    ///     DisplaySize128x64,
     ///     DisplayRotation::Rotate90,
     /// );
     /// assert_eq!(rotated_disp.get_dimensions(), (64, 128));
     /// ```
     pub fn get_dimensions(&self) -> (u8, u8) {
-        let (w, h) = self.display_size.dimensions();
-
         match self.display_rotation {
-            DisplayRotation::Rotate0 | DisplayRotation::Rotate180 => (w, h),
-            DisplayRotation::Rotate90 | DisplayRotation::Rotate270 => (h, w),
+            DisplayRotation::Rotate0 | DisplayRotation::Rotate180 => (DSIZE::WIDTH, DSIZE::HEIGHT),
+            DisplayRotation::Rotate90 | DisplayRotation::Rotate270 => (DSIZE::HEIGHT, DSIZE::WIDTH),
         }
     }
 
@@ -264,7 +242,7 @@ where
     }
 
     /// Change into any mode implementing DisplayModeTrait
-    pub fn into<NMODE: DisplayModeTrait<DI>>(self) -> NMODE
+    pub fn into<NMODE: DisplayModeTrait<DI, DSIZE>>(self) -> NMODE
     where
         DI: WriteOnlyDataCommand,
     {